feat(Brandon-UPS-002): reorganizar clases en paquetes y crear package controller con lista de productos
