+++ src/main/java/ec/edu/ups/models/Calculable.java
@@ -1,4 +1,4 @@
-package ec.edu.ups;
+package ec.edu.ups.models;
 
 public interface Calculable {
     double calcularCostoTotal();
