+++ src/main/java/ec/edu/ups/models/Departamento.java
@@ -1,4 +1,4 @@
-package ec.edu.ups;
+package ec.edu.ups.models;
 
 public enum Departamento {
     RECURSOS_HUMANOS,
