+++ src/main/java/ec/edu/ups/models/DetalleCompraPaquete.java
@@ -1,4 +1,4 @@
-package ec.edu.ups;
+package ec.edu.ups.models;
 
 import java.util.ArrayList;
 import java.util.List;
