+++ src/main/java/ec/edu/ups/models/UnidadMedida.java
@@ -1,4 +1,4 @@
-package ec.edu.ups;
+package ec.edu.ups.models;
 
 public enum UnidadMedida {
     METROS,
